requirements.txt
pytube3
ffmpeg-python